+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 History
 =======
 
+0.5.1 (June, 2019)
+-------------------------
+
+* Tokens are represented internally as String
+* Asset is registered first on-chain
+* Keeper sleep duration and retries can be configured
+* Updated to Brizo 0.3.9 & Aquarius 0.2.8
+
 
 0.5.0 (May 17th, 2019)
 -------------------------